requirements.txt
# dev and test
# ------------
pytest==7.4.3
pytest_mock==3.12.0

# Code linters, formatters, and security scanners
# ------------
black==23.12.0
flake8==6.1.0
flake8-coding==1.3.2
pre-commit==3.6.0
isort==5.13.2
mypy==1.7.1
pylint==3.0.3
bandit==1.7.6
pydocstringformatter==0.7.3
tox==4.11.4
codespell==2.2.6

# production
# ------------
python-decouple==3.8
langchain==0.0.351
langchainhub==0.1.14
langchain-experimental==0.0.47
openai==1.6.0
pinecone-client==2.2.4
pinecone-text==0.7.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pypdf==3.17.3
tiktoken==0.5.2
pyodbc==5.0.1
dropbox==11.36.2
bs4==0.0.1
streamlit==1.31.1
